Fixed role issue (see desc), changed context proc
Fixed role creation.   If the role.label attr is saved it breaks being able to associate the role with a user.  So I took it out.  I also modified the context processor so instead of is_admin its now render().  Also added javascript to layout that highlights link of currently requested page or subpages.